requirements.txt
boto==2.26.1
requests==2.2.1
argparse==1.2.1
boto==2.26.1
distribute==0.6.34
python-firebase==1.1
requests==2.2.1
wsgiref==0.1.2
